# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Memora important days, reminders, notifications, devices, auth, and tasks",
        "title": "Memora API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/devices": {
            "get": {
                "description": "List active Expo push tokens for the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "List devices",
                "operationId": "list-devices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.DeviceTokenList"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "description": "Register or reactivate an Expo push token for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Register device",
                "operationId": "register-device",
                "parameters": [
                    {
                        "description": "Device token data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.RegisterDevice"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.DeviceToken"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/devices/{id}": {
            "delete": {
                "description": "Deactivate a registered device token by ID",
                "tags": [
                    "devices"
                ],
                "summary": "Delete device",
                "operationId": "delete-device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device token ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/devices/{id}/test-push": {
            "post": {
                "description": "Send a test Expo push notification to a registered device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Send test push",
                "operationId": "test-push",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device token ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Test push copy",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/v1.TestPush"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.PushTestResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/mobile/bootstrap": {
            "get": {
                "description": "Get initial data needed by the mobile app after login",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mobile"
                ],
                "summary": "Get mobile bootstrap",
                "operationId": "mobile-bootstrap",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "Upcoming lookahead window in days",
                        "name": "upcoming_days",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "Upcoming important days limit",
                        "name": "upcoming_limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Upcoming important days offset",
                        "name": "upcoming_offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.MobileBootstrap"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "entity.DeviceToken": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "Pixel 8"
                },
                "platform": {
                    "type": "string",
                    "example": "android"
                },
                "token": {
                    "type": "string",
                    "example": "ExpoPushToken[xxxxxxxxxxxxxxxxxxxxxx]"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-01-01T00:00:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "entity.ImportantDayType": {
            "type": "string",
            "enum": [
                "birthday",
                "wedding",
                "memorial",
                "graduation",
                "first_day",
                "document",
                "subscription",
                "medical",
                "custom"
            ],
            "x-enum-varnames": [
                "ImportantDayTypeBirthday",
                "ImportantDayTypeWedding",
                "ImportantDayTypeMemorial",
                "ImportantDayTypeGraduation",
                "ImportantDayTypeFirstDay",
                "ImportantDayTypeDocument",
                "ImportantDayTypeSubscription",
                "ImportantDayTypeMedical",
                "ImportantDayTypeCustom"
            ]
        },
        "entity.ImportantDayUpcoming": {
            "type": "object",
            "properties": {
                "anniversary": {
                    "type": "integer",
                    "example": 56
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-01-01T00:00:00Z"
                },
                "days_until": {
                    "type": "integer",
                    "example": 7
                },
                "description": {
                    "type": "string",
                    "example": "Buy flowers"
                },
                "event_day": {
                    "type": "integer",
                    "example": 13
                },
                "event_month": {
                    "type": "integer",
                    "example": 5
                },
                "event_year": {
                    "type": "integer",
                    "example": 1970
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "occurrence_date": {
                    "type": "string",
                    "example": "2026-05-13"
                },
                "person_name": {
                    "type": "string",
                    "example": "Mom"
                },
                "recurrence": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Recurrence"
                        }
                    ],
                    "example": "yearly"
                },
                "relationship": {
                    "type": "string",
                    "example": "mother"
                },
                "reminder_time": {
                    "type": "string",
                    "example": "09:00"
                },
                "timezone": {
                    "type": "string",
                    "example": "Asia/Jakarta"
                },
                "title": {
                    "type": "string",
                    "example": "Mom birthday"
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.ImportantDayType"
                        }
                    ],
                    "example": "birthday"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-01-01T00:00:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "entity.PushTestResult": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "sent_at": {
                    "type": "string",
                    "example": "2026-01-01T00:00:00Z"
                },
                "ticket_id": {
                    "type": "string",
                    "example": "00000000-0000-0000-0000-000000000000"
                }
            }
        },
        "entity.Recurrence": {
            "type": "string",
            "enum": [
                "yearly"
            ],
            "x-enum-varnames": [
                "RecurrenceYearly"
            ]
        },
        "entity.ReminderChannel": {
            "type": "string",
            "enum": [
                "email",
                "in_app",
                "push"
            ],
            "x-enum-varnames": [
                "ReminderChannelEmail",
                "ReminderChannelInApp",
                "ReminderChannelPush"
            ]
        },
...
        "entity.UserSettings": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-01-01T00:00:00Z"
                },
                "notification_channels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ReminderChannel"
                    },
                    "example": [
                        "email"
                    ]
                },
                "reminder_time": {
                    "type": "string",
                    "example": "09:00"
                },
                "timezone": {
                    "type": "string",
                    "example": "Asia/Jakarta"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-01-01T00:00:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "v1.DeviceTokenList": {
            "type": "object",
            "properties": {
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.DeviceToken"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "v1.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
...
        "v1.MobileBootstrap": {
            "type": "object",
            "properties": {
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.DeviceToken"
                    }
                },
                "devices_total": {
                    "type": "integer",
                    "example": 1
                },
                "settings": {
                    "$ref": "#/definitions/entity.UserSettings"
                },
                "unread_notification_count": {
                    "type": "integer",
                    "example": 2
                },
                "upcoming_important_days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ImportantDayUpcoming"
                    }
                },
                "upcoming_total": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
        "v1.RegisterDevice": {
            "type": "object",
            "required": [
                "platform",
                "token"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Pixel 8"
                },
                "platform": {
                    "type": "string",
                    "maxLength": 40,
                    "example": "android"
                },
                "token": {
                    "type": "string",
                    "example": "ExpoPushToken[xxxxxxxxxxxxxxxxxxxxxx]"
                }
            }
        },
...
        "v1.TestPush": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Push notifications are working."
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "Memora test"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}